# openapi.json
{
  "components": {
    "schemas": {
      "AISettingsPayload": {
        "description": "Body of the ``ai`` object for PUT /settings and POST /settings/test.\n\n``{env.*}`` template strings are preserved verbatim (plain str type);\napi_key/base_url omitted means \"keep the stored value\" (handled in the\nroute, not here).",
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_key_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "default": "",
            "title": "Provider",
            "type": "string"
          }
        },
        "title": "AISettingsPayload",
        "type": "object"
      },
      "AISettingsView": {
        "properties": {
          "api_key_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env"
          },
          "api_key_set": {
            "title": "Api Key Set",
            "type": "boolean"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model",
          "api_key_set",
          "enabled"
        ],
        "title": "AISettingsView",
        "type": "object"
      },
      "AnswerRequest": {
        "properties": {
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "AnswerRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SettingsPutRequest": {
        "properties": {
          "ai": {
            "$ref": "#/components/schemas/AISettingsPayload"
          }
        },
        "required": [
          "ai"
        ],
        "title": "SettingsPutRequest",
        "type": "object"
      },
      "SettingsPutResponse": {
        "properties": {
          "config_path": {
            "title": "Config Path",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "config_path"
        ],
        "title": "SettingsPutResponse",
        "type": "object"
      },
      "SettingsResponse": {
        "properties": {
          "ai": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AISettingsView"
              },
              {
                "type": "null"
              }
            ]
          },
          "config_path": {
            "title": "Config Path",
            "type": "string"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          }
        },
        "required": [
          "configured",
          "config_path"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
      "SettingsTestRequest": {
        "properties": {
          "ai": {
            "$ref": "#/components/schemas/AISettingsPayload"
          }
        },
        "required": [
          "ai"
        ],
        "title": "SettingsTestRequest",
        "type": "object"
      },
      "SettingsTestResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "detail"
        ],
        "title": "SettingsTestResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "oxe",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/answer": {
      "post": {
        "description": "SSE-stream an AI answer for a query. Requires a configured model.",
        "operationId": "ai_answer_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Answer"
      }
    },
...
    "/settings": {
      "get": {
        "description": "Current [ai] config for the settings dialog; api_key redacted.",
        "operationId": "get_settings_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings"
      },
      "put": {
        "description": "Write the [ai] config section. api_key left untouched when omitted.",
        "operationId": "put_settings_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsPutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Settings"
      }
    },
    "/settings/test": {
      "post": {
        "description": "Verify provider + api_key + base_url + model with a minimal call.\n\nBody: {\"ai\": {provider, model, base_url?, api_key?}}; blank/omitted\napi_key falls back to the stored key (same as PUT /settings).\nReturns {ok: bool, detail: str}. Lazy imports, ~10s timeout.",
        "operationId": "test_settings_settings_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Settings"
      }
    },
...
  }
}